requirements.txt
pandas
numpy
deepface
opencv-python
retinaface
Pillow
tqdm
tensorflow
